App/requirements.txt
Django==5.1.5
gunicorn==21.2.0
whitenoise==6.7.0
djangorestframework==3.15.2
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
numpy==1.26.4
opencv-python-headless==4.11.0.86
pillow==11.0.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
sqlparse==0.5.3
tensorflow-cpu==2.17.1
tf_keras==2.17.0
keras==3.8.0
h5py==3.12.1
joblib==1.4.2
threadpoolctl==3.5.0
urllib3==2.3.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
